Use file:script if a .config.script file present
When reading rebar.config or reltool.config, rebar will check whether there
exists a file with the same name, but with the additional extension .script
(e.g. rebar.config.script, or reltool.config.script). If so, it will first
read the original .config file (if present) using file:consult/1, then
apply file:script(Script, Bindings) to the .script file. The contents of the
original .config file will be bound to the variable CONFIG, and the full name
of the .script file is bound to the variable SCRIPT.

rebar_rel_utils:is_rel_dir/1 will now return true if Dir contains a
reltool.config.script file or a reltool.config file. The return value is
{true, F}, where F is the found file (reltool.config.script if both were
present).